Working: 1 securityScheme, matching securedBy
Not happy with appending the security schemas parameter at the end, though. This should probably become the first parameter, and bind it in the renderHtml function on index, like is done with the baseUri. However, I thought changing all the getCurlStatement tests in the first go-round was too big a step